+++ src/main/kotlin/com/github/strindberg/emacsj/universal/UniversalArgumentDelegate.kt
@@ -6,6 +6,7 @@ import com.github.strindberg.emacsj.duplicate.ACTION_COPY_ABOVE_COMMAND
 import com.github.strindberg.emacsj.line.ACTION_TRANSPOSE_LINES
 import com.github.strindberg.emacsj.mark.ACTION_POP_MARK
 import com.github.strindberg.emacsj.mark.ACTION_PUSH_MARK
+import com.github.strindberg.emacsj.paste.ACTION_HISTORY_PASTE
 import com.github.strindberg.emacsj.paste.ACTION_PASTE
 import com.github.strindberg.emacsj.paste.ACTION_PREFIX_PASTE
 import com.github.strindberg.emacsj.search.ACTION_ISEARCH_BACKWARD
@@ -46,6 +47,7 @@ internal val singleActions = setOf(
     ACTION_REPLACE_REGEXP,
     ACTION_PASTE,
     ACTION_PREFIX_PASTE,
+    ACTION_HISTORY_PASTE,
     ACTION_PUSH_MARK,
     ACTION_POP_MARK,
     ACTION_ZAP_FORWARD_TO,