+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ## [Unreleased]
 
 ### Added
+
+- New Paste functionality: "Paste: Previous Item in Clipboard History" (`alt-y`) invoked on its own displays the kill ring as selectable
+  items. `ENTER` pastes the selected entry. 
 - When using Isearch, the search field now marks non-matching part of the search string in red.
 - When using Isearch, the kill ring can now be cycled with `alt-y/command-y` after pasting to search text.